Remove github refresh dependency, if-plugins dependencies
