+++ src/lib/gcf-regex/index.ts
@@ -7,7 +7,7 @@ import { validate } from '../../util/validations';
 
 import { PluginInterface } from '../../interfaces';
 import { ConfigParams, PluginParams } from '../../types/common';
-import caseMatch = require("./caseMatcher.cjs")
+import caseMatch from './caseMatcher'
 
 const { InputValidationError, ConfigValidationError } = ERRORS;
